Completed front end prior to integration
This has all the features advertised. Dynamically updating chart of connections, whitelist feature, speed control, and all the dragging, collapsing, and clicking functionality needed for base functionality. 

There's also a stub in visualFirewallController for where the pcap information goes, and once this is integrated with the actual JSONS, I'll add in the scaling link size based on bandwidth incoming.